Optimize SVG icon
